Develop

+++ apps/api/src/modules/position-request/position-request.service.ts
@@ -1,6 +1,7 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { Field, Int, ObjectType } from '@nestjs/graphql';
 import { Prisma } from '@prisma/client';
+import { JsonObject } from '@prisma/client/runtime/library';
 import { btoa } from 'buffer';
 import { Elements, autolayout, generateJobProfile } from 'common-kit';
 import dayjs from 'dayjs';
@@ -52,7 +53,7 @@ export class PositionRequestResponse {
   parent_job_profile_id: number;
 
   @Field(() => GraphQLJSON)
-  profile_json: any;
+  profile_json_updated: any;
 
   @Field(() => GraphQLJSON)
   orgchart_json: any;
@@ -131,11 +132,11 @@ export class PositionRequestApiService {
     return this.prisma.positionRequest.create({
       data: {
         department: data.department,
-        additional_info: data.additional_info,
+        additional_info: data.additional_info === null ? Prisma.DbNull : data.additional_info,
         step: data.step,
         reports_to_position_id: data.reports_to_position_id,
-        profile_json: data.profile_json,
-        orgchart_json: data.orgchart_json,
+        profile_json_updated: data.profile_json_updated === null ? Prisma.DbNull : data.profile_json_updated,
+        orgchart_json: data.orgchart_json === null ? Prisma.DbNull : data.orgchart_json,
         // TODO: AL-146
         // user: data.user,
         user_id: userId,
@@ -163,9 +164,8 @@ export class PositionRequestApiService {
         where: { id },
         data: {
           additional_info: {
-            update: {
-              comments: comment,
-            },
+            ...(positionRequest.additional_info as JsonObject),
+            comments: comment,
           },
         },
       });
@@ -353,6 +353,7 @@ export class PositionRequestApiService {
         approved_at: true,
         submitted_at: true,
         crm_id: true,
+        crm_lookup_name: true,
         shareUUID: true,
         additional_info: true,
       },
@@ -777,12 +778,13 @@ export class PositionRequestApiService {
       updatePayload.reports_to_position_id = updateData.reports_to_position_id;
     }
 
-    if (updateData.profile_json !== undefined) {
-      updatePayload.profile_json = updateData.profile_json;
+    if (updateData.profile_json_updated !== undefined) {
+      updatePayload.profile_json_updated =
+        updateData.profile_json_updated === null ? Prisma.DbNull : updateData.profile_json_updated;
     }
 
     if (updateData.orgchart_json !== undefined) {
-      updatePayload.orgchart_json = updateData.orgchart_json;
+      updatePayload.orgchart_json = updateData.orgchart_json === null ? Prisma.DbNull : updateData.orgchart_json;
     }
 
     if (updateData.title !== undefined) {
@@ -945,9 +947,9 @@ export class PositionRequestApiService {
       if (additionalInfo.comments !== undefined) {
         (updatePayload.additional_info as Record<string, Prisma.JsonValue>).comments = additionalInfo.comments;
       }
-    } else if (additionalInfo == null) {
+    } else if (additionalInfo === null) {
       updatingAdditionalInfo = true;
-      updatePayload.additional_info = null;
+      updatePayload.additional_info = Prisma.DbNull;
     }
 
     if (!updatingAdditionalInfo) delete updatePayload.additional_info;
@@ -1001,8 +1003,8 @@ export class PositionRequestApiService {
     }
   }
 
-  async deletePositionRequest(id: number) {
-    const result = await this.prisma.positionRequest.delete({ where: { id } });
+  async deletePositionRequest(id: number, userId: string) {
+    const result = await this.prisma.positionRequest.delete({ where: { id, user_id: userId, status: 'DRAFT' } });
     return result;
   }
 
@@ -1027,7 +1029,7 @@ export class PositionRequestApiService {
     }
 
     // If the job profile is _not_ denoted as requiring review, it must be reviewed _only_ if significant sections have been changed
-    const prJobProfile = positionRequest.profile_json as Record<string, any>;
+    const prJobProfile = positionRequest.profile_json_updated as Record<string, any>;
 
     // Find position request job profile signficant sections
     const prJobProfileSignificantSections = {
@@ -1093,7 +1095,8 @@ export class PositionRequestApiService {
       this.dataHasChanges(
         JSON.stringify(jobProfileSignficantSections.education),
         JSON.stringify(prJobProfileSignificantSections.education),
-      ) && jobProfile.jobFamilies.some((jf) => jf.jobFamily.name != 'Administrative Services'), // AL-619 this is a temporary measure to disable education requirements for admin family
+      ) && !jobProfile.jobFamilies.some((jf) => jf.jobFamily.name == 'Administrative Services'), // AL-619 this is a temporary measure to disable education requirements for admin family
+
       // Job Experience
       this.dataHasChanges(
         JSON.stringify(jobProfileSignficantSections.job_experience),
@@ -1153,9 +1156,9 @@ export class PositionRequestApiService {
     });
 
     const jobProfileDocument =
-      positionRequest.profile_json != null
+      positionRequest.profile_json_updated != null
         ? generateJobProfile({
-            jobProfile: positionRequest.profile_json as Record<string, any>,
+            jobProfile: positionRequest.profile_json_updated as Record<string, any>,
             parentJobProfile: parentJobProfile,
           })
         : null;
@@ -1288,6 +1291,15 @@ export class PositionRequestApiService {
       incident = await this.crmService.getIncident(positionRequest.crm_id);
     }
 
+    if (incident.lookupName != null) {
+      await this.prisma.positionRequest.update({
+        where: { id: positionRequest.id },
+        data: {
+          crm_lookup_name: incident.lookupName,
+        },
+      });
+    }
+
     return incident;
   }
 
@@ -1337,7 +1349,7 @@ export class PositionRequestApiService {
           DESCR: positionRequest.title,
           REG_TEMP: PositionDuration.Regular,
           FULL_PART_TIME: PositionType.FullTime,
-          TGB_E_CLASS: `P${(positionRequest.profile_json as Record<string, any>).number}`,
+          TGB_E_CLASS: `P${(positionRequest.profile_json_updated as Record<string, any>).number}`,
           TGB_APPRV_MGR: employeeId,
         };
